1_Hrnet2Onnx/requirements.txt
opencv-python
PyYAML
scipy
tqdm
onnx
onnxruntime
yacs
onnx-simplifier